refactor: Trim down package to just naive variants
Played around with a python implementation using index lists [1], which
must have been where I translated the original code, and it seems wrong;
if I try putting in "hey" and run in through `bwt` and `ibwt` with the
generated indexlist, it doesn't get back to "hey". I could never quite
get the inverse to work myself, and I haven't found a great paper
describing the index list method more thoroughly for now, so remove it
and just keep around the naive versions with an exposed EOFChar that can
be modified by callers.

[1]: https://learn.fmi.uni-sofia.bg/pluginfile.php/169284/mod_resource/content/1/Burrows---Wheeler_transform.pdf